Fix compiler warning on float to int cast.
git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk@5976 640d5228-2204-0410-8d71-8f46fa6850b4

+++ nimbus/src/amlib/std/plwcc.c
@@ -112,7 +112,7 @@ void plwccInit(var_base *varp)
     AddToDefaults(varp->name, defaultName, values);
   }
   else
-    tau_Nu[probe_index] = tmp[0];
+    tau_Nu[probe_index] = (int)tmp[0];
 
   if ((tmp = GetDefaultsValue("KING_CLOUD_CONC_THRESHOLD", varp->name)) == NULL)
   {
